chore: Disable Disqus
While I liked comments on the site, I finally decided to kill Disqus due to the long loading times and tracking.

Most pages never got a comment, and I missed those on the one that did. Redirecting people to email is probably better. Maybe Hyvor Talk could be an alternative in the future.